requirements.txt
tqdm
notebook
openai
elasticsearch
pandas
scikit-learn
ipywidgets
seaborn
python-dotenv
ipykernel
transformers
accelerate
bitsandbytes
sentencepiece
sentence_transformers
psycopg[c]
psycopg-binary
pgcli
importlib_metadata
datasets
librosa
soundfile
spacy